refactor(inventory): split workspace parser modules
