+++ .sampo/changesets/922-split-inventory-parser.md
@@ -1,0 +1,5 @@
+---
+npm/@wp-typia/project-tools: patch
+---
+
+Split workspace inventory parser descriptors, entry parsing, and validation helpers into focused runtime modules.
